# JAVA-899 Use isTraceEnabled in CodecRegistry where trace is used.