Minor change to use 'isTraceEnabled()' in CodecRegistry for [JAVA-899](https://datastax-oss.atlassian.net/browse/JAVA-899).  After some microbenchmarking this doesn't seem to have made any difference throughput-wise, but I still think it's worth checking.  There were 2 other trace calls in the driver in Connection.java that don't check isTraceEnabled that I will update on the 2.0 branch.

##### Without isTraceEnabled() #####
```
Benchmark                                                Mode  Cnt         Score        Error  Units
CodecRegistryBenchmark.codecForWithDataTypeAndClass     thrpt   40  10161877.234 ± 537363.267  ops/s
CodecRegistryBenchmark.codecForWithDataTypeAndToken     thrpt   40  10997595.418 ± 620316.110  ops/s
CodecRegistryBenchmark.codecForWithNoDataType           thrpt   40    856435.550 ±  43492.977  ops/s
CodecRegistryBenchmark.codecForWithNoDataTypeBestCase   thrpt   40   4286882.168 ± 191963.935  ops/s
CodecRegistryBenchmark.codecForWithNoDataTypeWorstCase  thrpt   40     68979.731 ±   3913.868  ops/s
RowBenchmark.getObject                                  thrpt   40   2528969.380 ± 172483.368  ops/s
RowBenchmark.getWithToken                               thrpt   40   2503465.997 ± 155421.694  ops/s
```

##### With isTraceEnabled() #####
```
Benchmark                                                Mode  Cnt         Score        Error  Units
CodecRegistryBenchmark.codecForWithDataTypeAndClass     thrpt   40   9593448.056 ± 430546.988  ops/s
CodecRegistryBenchmark.codecForWithDataTypeAndToken     thrpt   40  10965278.538 ± 523017.248  ops/s
CodecRegistryBenchmark.codecForWithNoDataType           thrpt   40    855637.588 ±  50067.211  ops/s
CodecRegistryBenchmark.codecForWithNoDataTypeBestCase   thrpt   40   4142871.211 ± 257956.479  ops/s
CodecRegistryBenchmark.codecForWithNoDataTypeWorstCase  thrpt   40     65764.214 ±   3059.630  ops/s
RowBenchmark.getObject                                  thrpt   40   2605139.928 ± 128551.449  ops/s
RowBenchmark.getWithToken                               thrpt   40   2411576.277 ± 113910.636  ops/s
```