@tolbertam 2015-08-26
Closing as this doesn't actually offer any benefit (see benchmark results and comments in the jira).